Patched bug in permutation test P-value estimation (Issue #16)
Previously the P-value was calculated from the number of null distribution values more extreme than the true value for a statistic. This leads to problems with sparse networks, where null distributions can have many observations with identical values (e.g. 0 for mean adjacency). In this case the P-value would still be quite small. This has been fixed so that P-values are calculated as the number of null distribution values as, or more, extreme than the true value for each statistic.