Fixed bugs in user input checking
Now handles a wider variety of inputs we'd expect to see for different types of analyses, and the checking process is abstracted out in a nicer way. This process also involved the abstraction of a number of pieces of code from the main routine, and a change to the user interface to 'modulePreservation'.