Restores previous OMP / BLAS state
The number of threads available to OpenMP and BLAS is now restored at the end of each function. Previously, the number of threads was restricted to 1 so as not to conflict with the multi-core parallelisation, but not restored, leading to the uninentional side effect of restricting BLAS / OpenMP in the user's future function calls.