Bugfix: proportion of module variables present
The reported proportion of variables in each module present in the test dataset is now reported correctly.

Previously, these ratios could be larger than 1. This was due to inconsistent sorting of module labels (one dataset as character labels, one as numeric) which lead to incorrect calculations of the returned 'propVarsPresent' where > 10 modules were present.

Fixes Issue #24