Improved contingency table
Contingency tables now reports the full size of each module as well as the number of variables present in the other dataset in the contigency table comparison. Modules where no variables are present in the other dataset are also included.

The requested 'modules' are now represented as characters internally. Previously, this cast was implicit when calling the internal C++ functions, so indice mismatch problems were thankfully avoided, outside of the returned contingency table.