Fixed bugs in data heatmap legends
Data heatmap legends in 'plotModule' and 'plotData' now only show the
range of values present in the data matrix, rather than showing a legend
that is symmetric around 0 in the case where the range of values in the
data spans both negative and positive values.

Axis ticks for the lowest and highest values on the heatmap legends
will now show a "<" or ">" sign respectively if the user manually
specifies a range of values to restrict the heatmap color palette to
that is smaller than the range of values that occurs in the data.