Bugfix: plots no longer crash on missing samples
Fixed bug in plotSquareHeatmap() that caused 'plotModule' and 'plotData' to crash when specifying 'orderSamplesBy' when not all samples from the 'orderSamplesBy' dataset were present in the dataset to be rendered. Similarly, this fixes bugs in 'plotCorrelation' and 'plotNetwork' arising when setting 'symmetric=TRUE' when nodes from the discovery network were not present in the test dataset to be rendered.