Samples present in only the test dataset now shown when plotting
Previously, when setting 'orderSamplesBy' only samples present in both
the 'orderSamplesBy' and 'test' datasets were shown. Now, samples
present in only the 'test' dataset are also shown on the data heatmap
and module summary bar charts after (below) the samples present in both
datasets are shown. A horizontal black line is rendered to denote the
two groups of samples.